requirements.txt
matplotlib==3.8.3
seaborn==0.13.2
numpy==1.24.3
Opencv-python==4.9.0.80
pandas==1.4.2
Pillow==9.4.0
torch==1.13.1
torchvision==0.14.1
easyocr==1.7.1
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.9
onnx==1.16.0
onnxruntime==1.17.1
tqdm